+++ haarpy/circular_ensembles.py
@@ -15,12 +15,14 @@
 Circular ensembles Python interface
 """
 
+from math import prod
 from fractions import Fraction
 from functools import lru_cache
 from typing import Union
 from collections import Counter
 from sympy import Symbol, Expr, fraction, factor, simplify
-from sympy.combinatorics import Permutation
+from sympy.combinatorics import Permutation, SymmetricGroup
+from sympy.core.numbers import Integer
 from haarpy import (
     weingarten_orthogonal,
     weingarten_symplectic,
@@ -103,6 +105,133 @@ def haar_integral_circular_orthogonal(
     )
 
     if isinstance(group_dimension, Expr):
+        numerator, denominator = fraction(simplify(integral))
+        integral = factor(numerator) / factor(denominator)
+
+    return integral
+
+
+@lru_cache
+def haar_integral_circular_symplectic(
+    sequences: tuple[tuple[Expr]], half_dimension: Expr
+) -> Expr:
+    """Returns integral over circular symplectic ensemble polynomial
+    sampled at random from the Haar measure
+
+    Args:
+        sequences (tuple[tuple[int]]) : Indices of matrix elements
+        half_dimension (Symbol) : Half the dimension of the unitary group
+
+    Returns:
+        Expr : Integral under the Haar measure
+
+    Raise:
+        ValueError : If sequences doesn't contain 2 tuples
+        ValueError : If tuples i and j are of odd size
+        TypeError: If dimension is int and sequence is not
+        TypeError: If the half_dimension is not int nor Symbol
+        ValueError: If all sequence indices are not between 0 and 2*dimension - 1
+        TypeError: If sequence containt something else than Expr
+        TypeError: If symbolic sequences have the wrong format
+    """
+    if len(sequences) != 2:
+        raise ValueError("Wrong sequence format")
+
+    seq_i, seq_j = sequences
+
+    degree = len(seq_i)
+
+    if degree % 2 or len(seq_j) % 2:
+        raise ValueError("Wrong sequence format")
+
+    if isinstance(half_dimension, int):
+        if not all(isinstance(i, int) for i in seq_i + seq_j):
+            raise TypeError
+        if not all(0 <= i <= 2 * half_dimension - 1 for i in seq_i + seq_j):
+            raise ValueError("The matrix indices are outside the dimension range")
+        if degree != len(seq_j):
+            return 0
+        coefficient = prod(
+            -1 if i < half_dimension else 1 for i in (seq_i + seq_j)[::2]
+        )
+        shifted_i = [
+            (
+                i + half_dimension
+                if i < half_dimension and index % 2 == 0
+                else i - half_dimension if index % 2 == 0 else i
+            )
+            for index, i in enumerate(seq_i)
+        ]
+        shifted_j = [
+            (
+                i + half_dimension
+                if i < half_dimension and index % 2 == 0
+                else i - half_dimension if index % 2 == 0 else i
+            )
+            for index, i in enumerate(seq_j)
+        ]
+
+    elif isinstance(half_dimension, Symbol):
+        if not all(isinstance(i, (int, Expr)) for i in seq_i + seq_j):
+            raise TypeError
+
+        if not all(
+            (
+                len(xpr.as_ordered_terms()) == 2
+                and xpr.as_ordered_terms()[0] == half_dimension
+                and isinstance(xpr.as_ordered_terms()[1], Integer)
+                and xpr.as_ordered_terms()[1] > 0
+            )
+            or xpr == half_dimension
+            for xpr in seq_i + seq_j
+            if isinstance(xpr, Expr)
+        ):
+            raise TypeError
+        if degree != len(seq_j):
+            return 0
+        coefficient = prod(
+            -1 if isinstance(i, int) else 1 for i in (seq_i + seq_j)[::2]
+        )
+        shifted_i = [
+            (
+                i + half_dimension
+                if isinstance(i, int) and index % 2 == 0
+                else (
+                    0
+                    if i == half_dimension and index % 2 == 0
+                    else i.as_ordered_terms()[1] if index % 2 == 0 else i
+                )
+            )
+            for index, i in enumerate(seq_i)
+        ]
+        shifted_j = [
+            (
+                i + half_dimension
+                if isinstance(i, int) and index % 2 == 0
+                else (
+                    0
+                    if i == half_dimension and index % 2 == 0
+                    else i.as_ordered_terms()[1] if index % 2 == 0 else i
+                )
+            )
+            for index, i in enumerate(seq_j)
+        ]
+
+    else:
+        raise TypeError
+
+    permutation_tuple = (
+        permutation
+        for permutation in SymmetricGroup(degree).generate()
+        if permutation(shifted_i) == shifted_j
+    )
+
+    integral = coefficient * sum(
+        weingarten_circular_symplectic(permutation, half_dimension)
+        for permutation in permutation_tuple
+    )
+
+    if isinstance(half_dimension, Expr):
         numerator, denominator = fraction(simplify(integral))
         integral = factor(numerator) / factor(denominator)