+++ haarpy/tests/test_circular_ensembles.py
@@ -16,181 +16,345 @@ Circular ensembles tests
 """
 
 from math import prod
+from random import randint
 from fractions import Fraction
 from sympy import Symbol, simplify, factorial, factorial2
 from sympy.combinatorics import SymmetricGroup
 import pytest
 import haarpy as ap
 
-d = Symbol('d')
+d = Symbol("d")
+
+# The values in the dictionary below were verified against Monte Carlo simulations
+cse_dict = {
+    ((0, 0), (0, 0), 1): 1,
+    ((0, 1), (0, 1), 1): 0,
+    ((0, 1, 2, 3), (0, 1, 2, 3), 2): 1 / 6,
+    ((0, 0, 0, 0), (0, 0, 0, 0), 2): 1 / 6,
+    ((0, 3, 0, 3), (0, 3, 0, 3), 2): 1 / 6,
+    ((0, 0, 0, 3), (0, 3, 0, 0), 2): 1 / 12,
+    ((0, 0, 0, 0, 0, 0), (0, 0, 0, 0, 0, 0), 3): 1 / 35,
+    ((5, 5, 5, 5, 5, 5), (5, 5, 5, 5, 5, 5), 3): 1 / 35,
+    ((2, 2, 5, 5, 5, 5), (2, 2, 5, 5, 5, 5), 3): 1 / 35,
+    ((2, 2, 5, 5, 5, 5), (2, 5, 5, 2, 5, 5), 3): 0,
+    ((2, 2, 5, 5, 3, 3), (2, 2, 5, 5, 3, 3), 3): 1 / 63,
+    ((7, 7, 0, 0, 0, 0, 0, 0), (7, 0, 0, 7, 0, 0, 0, 0), 4): -1 / 4200,
+}
 
 
-@pytest.mark.parametrize("half_degree", range(1,3))
+@pytest.mark.parametrize("half_degree", range(1, 3))
 def test_weingarten_circular_orthogonal_hyperoctahedral_symbolic(half_degree):
     """Symbolic validation of COE Weingarten function against results shown in
-    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces: 
+    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces:
     <https://arxiv.org/abs/1301.5401>`_
     """
     if half_degree == 1:
-        for permutation in SymmetricGroup(2*half_degree).generate():
-            assert ap.weingarten_circular_orthogonal(permutation, d) == 1/(d+1)
+        for permutation in SymmetricGroup(2 * half_degree).generate():
+            assert ap.weingarten_circular_orthogonal(permutation, d) == 1 / (d + 1)
     else:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             hyperoctahedral = ap.HyperoctahedralGroup(half_degree)
-            coefficient = 1/(d*(d+1)*(d+3))
+            coefficient = 1 / (d * (d + 1) * (d + 3))
             assert ap.weingarten_circular_orthogonal(permutation, d) == (
-                simplify((d+2)*coefficient) if permutation in hyperoctahedral
-                else -coefficient
+                simplify((d + 2) * coefficient) if permutation in hyperoctahedral else -coefficient
             )
 
 
-@pytest.mark.parametrize("half_degree", range(1,3))
+@pytest.mark.parametrize("half_degree", range(1, 3))
 def test_weingarten_circular_orthogonal_hyperoctahedral_numeric(half_degree):
     """Symbolic validation of COE Weingarten function against results shown in
-    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces: 
+    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces:
     <https://arxiv.org/abs/1301.5401>`_
     """
     if half_degree == 1:
-        for permutation in SymmetricGroup(2*half_degree).generate():
-            assert ap.weingarten_circular_orthogonal(permutation, 7) == 1/(7+1)
+        for permutation in SymmetricGroup(2 * half_degree).generate():
+            assert ap.weingarten_circular_orthogonal(permutation, 7) == 1 / (7 + 1)
     else:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             hyperoctahedral = ap.HyperoctahedralGroup(half_degree)
-            coefficient = Fraction(1,(7*(7+1)*(7+3)))
+            coefficient = Fraction(1, (7 * (7 + 1) * (7 + 3)))
             assert ap.weingarten_circular_orthogonal(permutation, 7) == (
-                simplify((7+2)*coefficient) if permutation in hyperoctahedral
-                else -coefficient
+                simplify((7 + 2) * coefficient) if permutation in hyperoctahedral else -coefficient
             )
 
 
-@pytest.mark.parametrize("half_degree", range(1,3))
+@pytest.mark.parametrize("half_degree", range(1, 3))
 def test_weingarten_circular_symplectic_hyperoctahedral_symbolic(half_degree):
     """Symbolic validation of CSE Weingarten function against results shown in
-    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces: 
+    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces:
     <https://arxiv.org/abs/1301.5401>`_
     """
     if half_degree == 1:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             assert ap.weingarten_circular_symplectic(permutation, d) == (
-                permutation.signature()/(2*d-1)
+                permutation.signature() / (2 * d - 1)
             )
     else:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             hyperoctahedral = ap.HyperoctahedralGroup(half_degree)
-            coefficient = permutation.signature()/(d*(2*d-1)*(2*d-3))
+            coefficient = permutation.signature() / (d * (2 * d - 1) * (2 * d - 3))
             assert ap.weingarten_circular_symplectic(permutation, d) == (
-                simplify((d-1)*coefficient) if permutation in hyperoctahedral
-                else coefficient/2
+                simplify((d - 1) * coefficient)
+                if permutation in hyperoctahedral
+                else coefficient / 2
             )
 
 
-@pytest.mark.parametrize("half_degree", range(1,3))
+@pytest.mark.parametrize("half_degree", range(1, 3))
 def test_weingarten_circular_symplectic_hyperoctahedral_numeric(half_degree):
     """Symbolic validation of CSE Weingarten function against results shown in
-    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces: 
+    `Matsumoto. Weingarten calculus for matrix ensembles associated with compact symmetric spaces:
     <https://arxiv.org/abs/1301.5401>`_
     """
     if half_degree == 1:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             assert ap.weingarten_circular_symplectic(permutation, 7) == Fraction(
                 permutation.signature(),
-                (2*7-1),
+                (2 * 7 - 1),
             )
     else:
-        for permutation in SymmetricGroup(2*half_degree).generate():
+        for permutation in SymmetricGroup(2 * half_degree).generate():
             hyperoctahedral = ap.HyperoctahedralGroup(half_degree)
-            coefficient = Fraction(permutation.signature(), (7*(2*7-1)*(2*7-3)))
+            coefficient = Fraction(permutation.signature(), (7 * (2 * 7 - 1) * (2 * 7 - 3)))
             assert ap.weingarten_circular_symplectic(permutation, 7) == (
-                simplify((7-1)*coefficient) if permutation in hyperoctahedral
-                else coefficient*Fraction(1,2)
+                simplify((7 - 1) * coefficient)
+                if permutation in hyperoctahedral
+                else coefficient * Fraction(1, 2)
             )
 
 
 @pytest.mark.parametrize(
     "sequences",
     [
-        ((0,1,2,3), (0,1)),
-        ((0,1,2,3,4,5), (0,1,2,3)),
-        ((0,1,2,3), (0,1,2,2)),
-        ((0,0,0,0), (0,0,0,1)),
-    ]
+        ((0, 1, 2, 3), (0, 1)),
+        ((0, 1, 2, 3, 4, 5), (0, 1, 2, 3)),
+        ((0, 1, 2, 3), (0, 1, 2, 2)),
+        ((0, 0, 0, 0), (0, 0, 0, 1)),
+    ],
 )
 def test_haar_integral_coe_zero(sequences):
     "test cases that return 0"
     assert not ap.haar_integral_circular_orthogonal(sequences, d)
 
 
-@pytest.mark.parametrize("power", range(2,8,2))
+@pytest.mark.parametrize("power", range(2, 8, 2))
 def test_haar_integral_coe_diagonal_entry(power):
     """Test Thm. 4.2 as seen in
     `Matsumoto. General moments of matrix elements from circular
     orthogonal ensembles <https://doi.org/10.1142/S2010326312500050>`_
     """
-    sequences = (power*(0,), power*(0,))
-    assert (
-        ap.haar_integral_circular_orthogonal(sequences, d)
-        == factorial2(power)/prod((d+i) for i in range(1,power,2))
+    sequences = (power * (0,), power * (0,))
+    assert ap.haar_integral_circular_orthogonal(sequences, d) == factorial2(power) / prod(
+        (d + i) for i in range(1, power, 2)
     )
 
 
-@pytest.mark.parametrize("power", range(2,8,2))
+@pytest.mark.parametrize("power", range(2, 8, 2))
 def test_haar_integral_coe_off_diagonal_entry(power):
     """Test Thm. 4.3 as seen in
     `Matsumoto. General moments of matrix elements from circular
     orthogonal ensembles <https://doi.org/10.1142/S2010326312500050>`_
     """
-    half_power = int(power/2)
-    sequences = (half_power*(0,1), half_power*(0,1))
-    assert (
-        ap.haar_integral_circular_orthogonal(sequences, d)
-        == factorial(half_power)/((d+(power-1))*prod((d+i) for i in range(half_power-1)))
+    half_power = int(power / 2)
+    sequences = (half_power * (0, 1), half_power * (0, 1))
+    assert ap.haar_integral_circular_orthogonal(sequences, d) == factorial(half_power) / (
+        (d + (power - 1)) * prod((d + i) for i in range(half_power - 1))
     )
 
 
-@pytest.mark.parametrize("power", range(2,6,2))
+@pytest.mark.parametrize("power", range(2, 6, 2))
 def test_haar_integral_coe_diagonal_entry_numeric(power):
     """Test Thm. 4.2 as seen in
     `Matsumoto. General moments of matrix elements from circular
     orthogonal ensembles <https://doi.org/10.1142/S2010326312500050>`_
     """
     dimension = 17
-    sequences = (power*(0,), power*(0,))
-    assert (
-        ap.haar_integral_circular_orthogonal(sequences, dimension)
-        == Fraction(factorial2(power),prod((dimension+i) for i in range(1,power,2)))
+    sequences = (power * (0,), power * (0,))
+    assert ap.haar_integral_circular_orthogonal(sequences, dimension) == Fraction(
+        factorial2(power), prod((dimension + i) for i in range(1, power, 2))
     )
 
 
-@pytest.mark.parametrize("power", range(2,6,2))
+@pytest.mark.parametrize("power", range(2, 6, 2))
 def test_haar_integral_coe_off_diagonal_entry_numeric(power):
     """Test Thm. 4.3 as seen in
     `Matsumoto. General moments of matrix elements from circular
     orthogonal ensembles <https://doi.org/10.1142/S2010326312500050>`_
     """
     dimension = 17
-    half_power = int(power/2)
-    sequences = (half_power*(0,1), half_power*(0,1))
-    assert (
-        ap.haar_integral_circular_orthogonal(sequences, dimension)
-        == Fraction(
-            factorial(half_power),
-            ((dimension+(power-1))*prod((dimension+i) for i in range(half_power-1))),
-        )
+    half_power = int(power / 2)
+    sequences = (half_power * (0, 1), half_power * (0, 1))
+    assert ap.haar_integral_circular_orthogonal(sequences, dimension) == Fraction(
+        factorial(half_power),
+        ((dimension + (power - 1)) * prod((dimension + i) for i in range(half_power - 1))),
     )
+
 
 @pytest.mark.parametrize(
     "sequences",
     [
         ((1,),),
-        ((1,2,3),),
-        ((1,2),(3,4),(4,5)),
+        ((1, 2, 3),),
+        ((1, 2), (3, 4), (4, 5)),
         "str",
-        ((1,2),(3,4,5)),
-        ((1,2,3),(3,4,5,6)),
-        ((1,2,3),(1,2,3)),
-    ]
+        ((1, 2), (3, 4, 5)),
+        ((1, 2, 3), (3, 4, 5, 6)),
+        ((1, 2, 3), (1, 2, 3)),
+    ],
 )
 def test_haar_integral_coe_value_error(sequences):
     "Test haar integral value error"
     with pytest.raises(ValueError, match="Wrong tuple format"):
         ap.haar_integral_circular_orthogonal(sequences, d)
+
+
+@pytest.mark.parametrize(
+    "seq_i, seq_j, half_dim",
+    [
+        ((0, 0), (0, 0), 1),
+        ((0, 1), (0, 1), 1),
+        ((0, 1, 2, 3), (0, 1, 2, 3), 2),
+        ((0, 0, 0, 0), (0, 0, 0, 0), 2),
+        ((0, 3, 0, 3), (0, 3, 0, 3), 2),
+        ((0, 0, 0, 3), (0, 3, 0, 0), 2),
+        ((0, 0, 0, 0, 0, 0), (0, 0, 0, 0, 0, 0), 3),
+        ((5, 5, 5, 5, 5, 5), (5, 5, 5, 5, 5, 5), 3),
+        ((2, 2, 5, 5, 5, 5), (2, 2, 5, 5, 5, 5), 3),
+        ((2, 2, 5, 5, 5, 5), (2, 5, 5, 2, 5, 5), 3),
+        ((2, 2, 5, 5, 3, 3), (2, 2, 5, 5, 3, 3), 3),
+        ((7, 7, 0, 0, 0, 0, 0, 0), (7, 0, 0, 7, 0, 0, 0, 0), 4),
+    ],
+)
+def test_haar_integral_circular_symplectic_monte_carlo_numeric(seq_i, seq_j, half_dim):
+    "Test haar integral circular symplectic moments against Monte Carlo simulation numeric"
+    integral = ap.haar_integral_circular_symplectic((seq_i, seq_j), half_dim)
+
+    mc_integral = cse_dict[(seq_i, seq_j, half_dim)]
+
+    assert float(integral) == mc_integral
+
+
+@pytest.mark.parametrize(
+    "seq_i, seq_j, half_dim",
+    [
+        ((0, 0), (0, 0), 1),
+        ((0, d), (0, d), 1),
+        ((0, 1, d, d + 1), (0, 1, d, d + 1), 2),
+        ((0, 0, 0, 0), (0, 0, 0, 0), 2),
+        ((0, 1 + d, 0, 1 + d), (0, 1 + d, 0, 1 + d), 2),
+        ((0, 0, 0, 1 + d), (0, 1 + d, 0, 0), 2),
+        ((0, 0, 0, 0, 0, 0), (0, 0, 0, 0, 0, 0), 3),
+        ((d + 2, d + 2, d + 2, d + 2, d + 2, d + 2), (d + 2, d + 2, d + 2, d + 2, d + 2, d + 2), 3),
+        ((2, 2, d + 2, d + 2, d + 2, d + 2), (2, 2, d + 2, d + 2, d + 2, d + 2), 3),
+        ((2, 2, d + 2, d + 2, d + 2, d + 2), (2, d + 2, d + 2, 2, d + 2, d + 2), 3),
+        ((2, 2, d + 2, d + 2, d, d), (2, 2, d + 2, d + 2, d, d), 3),
+        ((d + 3, d + 3, 0, 0, 0, 0, 0, 0), (d + 3, 0, 0, d + 3, 0, 0, 0, 0), 4),
+    ],
+)
+def test_haar_integral_circular_symplectic_monte_carlo_symbolic(seq_i, seq_j, half_dim):
+    "Test haar integral circular symplectic moments against Monte Carlo simulation symbolic"
+    integral = ap.haar_integral_circular_symplectic((seq_i, seq_j), d).subs(d, half_dim)
+
+    seq_i = tuple(i if isinstance(i, int) else i.subs(d, half_dim) for i in seq_i)
+    seq_j = tuple(i if isinstance(i, int) else i.subs(d, half_dim) for i in seq_j)
+
+    mc_integral = cse_dict[((seq_i, seq_j, half_dim))]
+
+    assert float(integral) == mc_integral
+
+
+
+@pytest.mark.parametrize(
+    "sequences",
+    [
+        ((1, 2, 3, 4), (1, 2, 3, 4), (1, 2, 3, 4)),
+        ((1, 2, 3, 4),),
+        ((1, 2, 3, 4, 5), (1, 2, 3, 4, 5)),
+        ((1, 2, 3, 4, 5, 6), (1, 2, 3, 4, 5, 6, 7)),
+    ],
+)
+def test_haar_integral_circular_symplectic_value_error_wrong_tuple(sequences):
+    "Value error for wrong sequence format"
+    with pytest.raises(ValueError, match="Wrong sequence format"):
+        ap.haar_integral_circular_symplectic(sequences, d)
+
+
+@pytest.mark.parametrize(
+    "sequences",
+    [
+        (("a", "b", "c", "d"), (1, 2, 3, 4)),
+        ((1, 1, d + 1, d + 1), (1, 1, 1, 1)),
+    ],
+)
+def test_haar_integral_circular_symplectic_type_error_integer_dimension(sequences):
+    "Type error for integer dimension with not integer sequences"
+    dimension = randint(1, 99)
+    with pytest.raises(TypeError):
+        ap.haar_integral_circular_symplectic(sequences, dimension)
+
+
+@pytest.mark.parametrize(
+    "sequences, dimension",
+    [
+        (((1, 3), (1, 3)), 1),
+        (((1, 2, 3, 5), (1, 2, 3, 4)), 2),
+        (((1, 2, 3, 41), (1, 2, 3, 41)), 20),
+    ],
+)
+def test_haar_integral_circular_symplectic_value_error_outside_dimension_range(
+    sequences, dimension
+):
+    "Value error for sequences values outside dimension range"
+    with pytest.raises(
+        ValueError,
+        match="The matrix indices are outside the dimension range",
+    ):
+        ap.haar_integral_circular_symplectic(sequences, dimension)
+
+
+@pytest.mark.parametrize(
+    "sequences",
+    [
+        ((1, 2, 3, 4), (1, 2, 3, "a")),
+        ((1, 2, 3, 4), (1, 2, 3, {1, 2})),
+        ((1, 2, 3, 4), (1, 2, 3, 4 * d)),
+        ((1, 2, 3, 2 * d + 1), (1, 2, 3, 4)),
+        ((1, 2, 3, d + 1), (1, 2, 3, 4.0)),
+        ((1, 2, 3, d - 1), (1, 2, 3, 4)),
+        ((1, 2, 3, 4), (1, 2, 3, d**2)),
+        ((1, 2, 3, 4), (1, 2, 3, 1 + d**2 + d)),
+        ((1, 2, 3, 4), (1, 2, 3, d + Symbol("s"))),
+    ],
+)
+def test_haar_integral_circular_symplectic_type_error_wrong_format(sequences):
+    "Type error for symbolic dimension with wrong sequence format"
+    with pytest.raises(TypeError):
+        ap.haar_integral_circular_symplectic(sequences, d)
+
+
+@pytest.mark.parametrize(
+    "dimension",
+    [
+        "a",
+        [1, 2],
+        {1, 2},
+        3.0,
+    ],
+)
+def test_haar_integral_circular_symplectic_wrong_dimension_format(dimension):
+    "Type error if the symplectic dimension is not an int nor a symbol"
+    with pytest.raises(TypeError):
+        ap.haar_integral_circular_symplectic(((1, 2, 3, 4), (1, 2, 3, 4)), dimension)
+
+
+@pytest.mark.parametrize(
+    "sequences, dimension",
+    [
+        (((1, 1), (1, 1, 1, 1)), d),
+        (((1, 1, d + 1, d + 2), (1, 1)), d),
+        (((0, 0, 0, 0), (0, 0, 0, 0, 2, 2)), 2),
+    ],
+)
+def test_haar_integral_circular_symplectic_zero_cases(sequences, dimension):
+    "Test cases that yield zero"
+    assert not ap.haar_integral_circular_symplectic(sequences, dimension)